[Style] 베스트상품 컴포넌트 변경 및 Card 컴포넌트 style 수정 #212
