+++ src/app/page.tsx
@@ -5,6 +5,7 @@ import { ProductDetail } from "./detail/[id]/page";
 import Banner from "./Banner";
 import Footer from "@/components/layout/Footer";
 import { Metadata } from "next";
+import Card from "@/components/Card";
 
 export const metadata: Metadata = {
   title: "조지주 메인",
@@ -35,7 +36,7 @@ async function fetchProductList(params?: string[][]): Promise<ProductDetail[]> {
 export default async function Home() {
   const bestProduct = await fetchProductList([
     ["sort", '{"buyQuantity": -1}'],
-    ["limit", "5"],
+    ["limit", "4"],
   ]);
   const newSortProduct = await fetchProductList([
     ["sort", '{ "createdAt": -1 }'],
@@ -64,13 +65,14 @@ export default async function Home() {
       <main className="px-0 bg-white">
         <CardSwiper data={newSortProduct} />
         <section id="product-best" className="mb-12 px-[25px]">
-          <h2 className="mt-5 mb-6 text-black subTitleMedium">베스트 상품</h2>
-          <ul className="mb-3 [&>*:last-child]:mb-0 ">
+          <h2 className="mt-5 text-black subTitleMedium">베스트 상품</h2>
+          <p className="mb-6 text-gray">최근 판매량이 가장 많은 상품이예요.</p>
+          <ul className="flex flex-row flex-wrap gap-3 mb-3">
             {bestProduct &&
               bestProduct.map((item) => {
                 return (
-                  <li className="mb-3" key={item._id}>
-                    <CardItemSmall data={item} />
+                  <li className="mb-3 w-[calc(50%-6px)]" key={item._id}>
+                    <Card data={item} />
                   </li>
                 );
               })}
