+++ src/components/Card.tsx
@@ -6,19 +6,20 @@ export default function Card({ data }: { data: ProductDetail }) {
   const API_SERVER = process.env.NEXT_PUBLIC_API_SERVER;
   return (
     <Link href={`/detail/${data._id}`}>
-      <div className="h-[240px] flex flex-col relative">
+      <div className="relative pb-[100%] mb-3">
         <Image
           src={API_SERVER + data.mainImages[0].path}
           alt={data.name}
-          width={182}
-          height={180}
-          className="round mb-3 h-full overflow-hidden"
+          className="round overflow-hidden"
+          fill
         />
-        <span className="contentMedium text-[black] text-ellipsis line-clamp-1 overflow-hidden">
+      </div>
+      <div>
+        <h3 className="contentMedium text-[black] text-ellipsis line-clamp-1 overflow-hidden">
           {data.name}
-        </span>
+        </h3>
         <span className="description text-[gray] pt-1">{data.extra.brewery}</span>
-        <span className="contentMedium text-[black] ">{data.price.toLocaleString()}원</span>
+        <p className="contentMedium text-[black] ">{data.price.toLocaleString()}원</p>
       </div>
     </Link>
   );
